+++ sdk/identity/azure-identity/inc/azure/identity/workload_identity_credential.hpp
@@ -8,6 +8,7 @@
 
 #pragma once
 
+#include "azure/identity/client_assertion_credential.hpp"
 #include "azure/identity/detail/client_credential_core.hpp"
 #include "azure/identity/detail/token_cache.hpp"
 
@@ -18,7 +19,7 @@
 
 namespace Azure { namespace Identity {
   namespace _detail {
-    class TokenCredentialImpl;
+    class ClientAssertionCredentialImpl;
   } // namespace _detail
 
   /**
@@ -74,11 +75,10 @@ namespace Azure { namespace Identity {
    */
   class WorkloadIdentityCredential final : public Core::Credentials::TokenCredential {
   private:
-    _detail::TokenCache m_tokenCache;
-    _detail::ClientCredentialCore m_clientCredentialCore;
-    std::unique_ptr<_detail::TokenCredentialImpl> m_tokenCredentialImpl;
-    std::string m_requestBody;
+    std::unique_ptr<_detail::ClientAssertionCredentialImpl> m_clientAssertionCredentialImpl;
     std::string m_tokenFilePath;
+
+    std::string GetAssertion(Core::Context const& context) const;
 
   public:
     /**