+++ sdk/keyvault/azure-security-keyvault-administration/inc/azure/keyvault/administration/backup_client.hpp
@@ -32,7 +32,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Administra
      * @brief Service Version used.
      *
      */
-    const std::string ApiVersion{"7.5"};
+    std::string ApiVersion{"7.5"};
   };
 
   /**