+++ sdk/keyvault/azure-security-keyvault-secrets/test/ut/secret_client_test.cpp
@@ -184,7 +184,33 @@ TEST_F(KeyVaultSecretClientTest, BackupRestore)
   {
     auto purgedResponse = client.PurgeDeletedSecret(secretName);
     CheckValidResponse(purgedResponse, Azure::Core::Http::HttpStatusCode::NoContent);
-    TestSleep(4min);
+    TestSleep(m_defaultWait);
+  }
+  {
+    int retries = 15;
+    // before we restore we need to ensure the secret is purged.
+    // since we have no visibility into the purge status we will just wait and check for a few
+    // seconds
+    while (retries > 15)
+    {
+      try
+      {
+        client.GetDeletedSecret(secretName);
+      }
+      catch (Azure::Core::RequestFailedException const&)
+      {
+        std::cout << std::endl << "- Secret is gone";
+        break;
+      }
+      TestSleep(m_defaultWait);
+      retries--;
+    }
+
+    if (retries == 0)
+    {
+      std::cout << std::endl << "retries reached 0";
+      EXPECT_TRUE(false);
+    }
   }
   {
     auto restore = client.RestoreSecretBackup(backupData);