+++ sdk/keyvault/azure-security-keyvault-administration/CHANGELOG.md
@@ -1,16 +1,20 @@
 # Release History
 
-## 4.0.0-beta.5 (Unreleased)
+## 4.0.0-beta.6 (Unreleased)
 
 ### Features Added
-
-- Add support for Backup/Restore operations for Key Vault HSM.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 4.0.0-beta.5 (2024-08-06)
+
+### Features Added
+
+- Add support for Backup/Restore operations for Key Vault HSM.
 
 ## 4.0.0-beta.4 (2024-01-16)