+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -1,12 +1,20 @@
 # Release History
 
-## 12.13.0-beta.1 (Unreleased)
+## 12.13.0-beta.2 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 12.13.0-beta.1 (2024-08-07)
+
+### Features Added
+
+- Added ability to retrieve SAS string to sign for debugging purposes.
 
 ### Other Changes