+++ sdk/storage/azure-storage-common/CHANGELOG.md
@@ -1,12 +1,21 @@
 # Release History
 
-## 12.8.0-beta.1 (Unreleased)
+## 12.8.0-beta.2 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 12.8.0-beta.1 (2024-08-07)
+
+### Features Added
+
+- Bumped up Account SAS version to `2024-11-04`.
+- Added ability to retrieve SAS string to sign for debugging purposes.
 
 ### Other Changes