+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.11.0-beta.1 (Unreleased)
+## 12.11.0-beta.2 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,16 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 12.11.0-beta.1 (2024-08-07)
+
+### Features Added
+
+- Bumped up API version to `2024-11-04`.
+- Added support for token-based authentication for all APIs.
+- Added support for paid bursting on premium file share accounts.
+- Added support for binary format for file permissions.
+- Added ability to retrieve SAS string to sign for debugging purposes.
 
 ## 12.10.0 (2024-07-16)