+++ sdk/tables/azure-data-tables/CHANGELOG.md
@@ -1,10 +1,16 @@
 # Release History
 
-## 1.0.0-beta.4 (Unreleased)
+## 1.0.0-beta.5 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0-beta.4 (2024-08-06)
 
 ### Bugs Fixed
 
@@ -15,8 +21,6 @@
 Thank you to our developer community members who helped to make Azure Data Tables better with their contributions to this release:
 
 - arwell _([GitHub](https://github.com/0xar1))_
-
-### Other Changes
 
 ## 1.0.0-beta.3 (2024-06-11)